backend/requirements.txt
# Saheli-AI Backend Dependencies
# ─────────────────────────────────────────────
# Core
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.7.0

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0           # Async PostgreSQL driver
alembic==1.14.0           # Database migrations

# File handling
python-multipart==0.0.20  # File uploads

# HTTP
httpx==0.28.1

# AI / LLM
openai==1.58.1
anthropic==0.40.0         # Optional: Claude support

# OCR
pytesseract==0.3.13       # Tesseract wrapper
Pillow==11.1.0            # Image processing
google-cloud-vision==3.8.2  # Google Cloud Vision (optional, production)

# Scheduler
APScheduler==3.11.0

# Utilities
python-dotenv==1.0.1
